Fix setting segmentation labelmap conversion color table
The segmentation labelmap conversion color table ID was incorrect and a new color table was created each time a segmentation was completed.

Also fixed error when running segmentation on a volume node that does not have a storage node (e.g., volumes that are loaded from DICOM or loaded using SampleData module).